+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## vN.N.N
 * Updating blocksmith sidebar to allow for complete customization of buttons.
+* Removed advanced copy and paste feature.
 
 
 ## v0.17.0