+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## vN.N.N
 * Updating blocksmith sidebar to allow for complete customization of buttons.
-* Removed advanced copy and paste feature.
+* Removed advanced copy and paste for mouse and keyboard.
 
 
 ## v0.17.0